+++ app/services/article_send_api.rb
@@ -24,7 +24,7 @@ class ArticleSendApi
       response = HTTParty.get("https://extractorapi.com/api/v1/extractor/?apikey=#{ENV['extractor_key']}&url=#{@url}&fields=domain,title,date_published,images,videos,clean_html,html")
       response_hash = JSON.parse(response.body)
 
-      raise NoMethodError if response_hash['text'].blank?
+      raise 'Response Fail, No Text!' if response_hash['text'].blank?
       @short_description = response_hash['text'].split('').first(50).join('')
       
       #萃取出 og:image 圖片位址
@@ -51,7 +51,7 @@ class ArticleSendApi
       obj.perform
     end
     rescue 
-      { other_status: 'extractor_fail' }
+      { api_status: 'extractor_fail' }
   end
